fix: ai review
